building out executor estimate

+++ pkg/internal/common/cost.go
@@ -1,13 +1,13 @@
 package common
 
-func NativeTokenBuffer(chainID uint32) float32 {
+func NativeTokenBuffer(chainID uint64) float64 {
 	return 0.05
 }
 
-func GasBuffer(chainID uint32) float32 {
+func GasBuffer(chainID uint64) float64 {
 	return 0.05
 }
 
-func TokenBuffer(token string) float32 {
+func TokenBuffer(token string) float64 {
 	return 0.05
 }
